+++ cli/types/index.d.ts
@@ -9,7 +9,6 @@
 
 /// <reference path="./cy-blob-util.d.ts" />
 /// <reference path="./cy-bluebird.d.ts" />
-/// <reference path="./cy-moment.d.ts" />
 /// <reference path="./cy-minimatch.d.ts" />
 /// <reference path="./cy-chai.d.ts" />
 /// <reference path="./lodash/index.d.ts" />