+++ packages/server/test/e2e/5_headless_spec.ts
@@ -40,10 +40,9 @@ describe('e2e headless', function () {
     })
   })
 
-  e2e.it('launches maximized by default in headless mode (1280x720)', {
-    browser: 'chrome',
+  e2e.it('launches maximized by default in headless mode (1920x1080)', {
     headed: false,
     project: Fixtures.projectPath('screen-size'),
-    spec: '720p.spec.js',
+    spec: 'default_size.spec.js',
   })
 })