+++ packages/server/test/integration/cypress_spec.js
@@ -1143,8 +1143,12 @@ describe('lib/cypress', () => {
             setProxy: sinon.stub().resolves(),
             setUserAgent: sinon.stub(),
             on: sinon.stub(),
+            removeListener: sinon.stub(),
           },
         }
+
+        ee.maximize = sinon.stub
+        ee.setSize = sinon.stub
 
         sinon.stub(browserUtils, 'launch').resolves(ee)
         sinon.stub(Windows, 'create').returns(ee)
@@ -1160,6 +1164,8 @@ describe('lib/cypress', () => {
             on: sinon.stub(),
             send: sinon.stub(),
           }
+
+          sinon.stub(chromeBrowser, '_writeExtension').resolves()
 
           sinon.stub(chromeBrowser, '_connectToChromeRemoteInterface').resolves(criClient)
           // the "returns(resolves)" stub is due to curried method